Moving util tests to test_utils
